// src/app/shared/model/app-other-info.model.ts
export class AppOtherInfoObj {
    AppOtherInfoId: number;
    AppId: number;
    BankCode: string;
    BankBranch: string;
    BankAccNo: string;
    BankAccName: string;
    AdditionalInterestPaidBy: string;
    RowVersion: string;
    constructor() { this.AppOtherInfoId = 0, this.RowVersion = ""; }
}